refactor: reorganize agentic AI connectors package structure
