+++ connectors/agentic-ai/agentic-ai-connectors/element-templates/hybrid/agenticai-aiagent-outbound-connector-hybrid.json
@@ -6,7 +6,6 @@
   "metadata" : {
     "keywords" : [ ]
   },
-  "documentationRef" : "https://example.com",
   "version" : 1,
   "category" : {
     "id" : "connectors",
