+++ connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/resolver/AdHocToolsSchema.java
@@ -4,8 +4,9 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema;
+package io.camunda.connector.agenticai.adhoctoolsschema.resolver;
 
-public interface AdHocToolsSchemaResolver {
-  AdHocToolsSchema resolveSchema(Long processDefinitionKey, String adHocSubprocessId);
-}
+import io.modelcontextprotocol.spec.McpSchema;
+import java.util.List;
+
+public record AdHocToolsSchema(List<McpSchema.Tool> tools) {}
