+++ connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/resolver/AdHocToolsSchemaResolver.java
@@ -4,9 +4,8 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema;
+package io.camunda.connector.agenticai.adhoctoolsschema.resolver;
 
-import io.modelcontextprotocol.spec.McpSchema;
-import java.util.List;
-
-public record AdHocToolsSchema(List<McpSchema.Tool> tools) {}
+public interface AdHocToolsSchemaResolver {
+  AdHocToolsSchema resolveSchema(Long processDefinitionKey, String adHocSubprocessId);
+}
